flying like a champ. now complete working integer control, shouldn't be a float anywhere onboard.

+++ drivers/pid.h
@@ -13,13 +13,13 @@
 typedef struct {
 	signed short p, i, d;
 	unsigned short i_limit; //total limit on the effect on I, set to 0 to disable.
-	signed int error;
+	signed long error;
 	signed short prev_val;
 }PID_DATA;
 
 
 //int16_t calculate_pid(int16_t incoming_val, int16_t goal_val, PID_DATA*pid);
-signed short calculate_pid(signed short incoming_val, signed short goal_val, PID_DATA*pid);
+signed int calculate_pid(signed short incoming_val, signed short goal_val, PID_DATA*pid);
 
 
 #endif /* PID_H_ */
